+++ np.Templating/lib/support/modules/FrontmatterModule.js
@@ -22,9 +22,12 @@ export function getBody(templateData: string = ''): string {
 }
 
 export default class FrontmatterModule {
-  constructor(config: any = {}) {
-    // $FlowFixMe
-    this.config = config
+  // $FlowIgnore
+  constructor(NPTemplating: any = null) {
+    if (NPTemplating) {
+      // $FlowIgnore
+      this.templatingInstance = NPTemplating
+    }
   }
 
   isFrontmatterTemplate(templateData: string): boolean {
@@ -43,13 +46,17 @@ export default class FrontmatterModule {
     return ''
   }
 
-  render(template: any = ''): any {
-    const fmData = fm(template)
-    Object.keys(fmData?.attributes).forEach((key) => {
-      fmData.attributes[key] ? fmData.attributes[key] : (fmData.attributes[key] = '')
-    })
+  parse(template: string = ''): any {
+    if (this.isFrontmatterTemplate(template)) {
+      const fmData = fm(template)
+      Object.keys(fmData?.attributes).forEach((key) => {
+        fmData.attributes[key] ? fmData.attributes[key] : (fmData.attributes[key] = '')
+      })
 
-    return fmData
+      return fmData
+    } else {
+      return {}
+    }
   }
 
   attributes(templateData: string = ''): any {